feat: dual variants for AudioChannelOps; usize channel indices
AudioChannelOps follows the dual-variant convention and channel indices are
now usize (counts keep ChannelCount/NonZeroU32). to_mono, to_stereo, and
duplicate_to_channels gain _in_place twins; swap_channels, pan, balance, and
apply_to_channel are renamed to *_in_place with non-mutating defaults.
extract_channel, borrow_channel, interleave, and deinterleave are unchanged
apart from the usize index arguments. Equivalence tests added for swap_channels
and to_mono.

+++ src/operations/editing.rs
@@ -1967,7 +1967,7 @@ where
             let mut shifted_channels = Vec::with_capacity(num_channels);
 
             for ch_idx in 0..num_channels {
-                let channel_audio = audio_owned.extract_channel(ch_idx as u32)?;
+                let channel_audio = audio_owned.extract_channel(ch_idx)?;
                 let shifted_channel = phase_vocoder_pitch_shift(
                     &channel_audio,
                     pitch_ratio,
